+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Video Manager and Audio Manager objects register themselves with a video player on startup
 - Updated sync video player integrity check to look for none or duplicate managers referencing it
 - Added menu items to GameObject menu to add different VideoTXL objects to scene
+- Added repair prefab instance tool to fix some classes of problems
+- Added URL Remapper prefab
 
 ## [2.0.0] - 07-29-23